backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.3.0
django-ckeditor==6.7.1
python-decouple==3.8
whitenoise==6.8.2
gunicorn==23.0.0